+++ doc/changelog.rst
@@ -23,16 +23,18 @@ General
   limit, are recolored by using the :ref:`constraint impedance<soParameters>` :math:`d` to mix the existing color with
   :ref:`visual/rgba/constraint<visual-rgba-constraint>`. For spatial tendons, this visualization aid is active only if
   no :ref:`material<tendon-spatial-material>` is set and :ref:`rgba<tendon-spatial-rgba>` is default.
-
 - Added :ref:`mju_getXMLDependencies` for computing a list of unique asset dependencies from an MJCF file.
-
 - Added the code sample ``dependencies`` which provides command line utility for printing the result of :ref:`mju_getXMLDependencies`.
 
+.. admonition:: Breaking ABI changes
+   :class: attention
+
+   - The attribute ``mjOption.apirate`` was unused and has been removed.
 
 MJX
 ^^^
-
 - Fix :github:issue:`2508`, ``qLD`` shapes mismatched mjModel during ``get_data_into``.
+
 
 Version 3.3.6 (September 15, 2025)
 ----------------------------------